+++ packages/sources/src/components/form-buttons/form-buttons.models.ts
@@ -1,0 +1,6 @@
+import { Button } from "../button/button.models";
+
+export interface FormButtons {
+  buttons: Button[];
+  asideButtons: Button[];
+}